+++ tests/analyses/test_wavefront.py
@@ -44,10 +44,32 @@ class TestZernikeStandardCoefficients:
         )
 
     @pytest.mark.parametrize("sampling,maximum_term", [("64x64", 37), ("128x128", 64)])
+    def test_zernike_standard_coefficients_decentered_returns_correct_result(
+        self, decentered_system, sampling, maximum_term, expected_data
+    ):
+        result = ZernikeStandardCoefficients(sampling=sampling, maximum_term=maximum_term).run(decentered_system)
+
+        assert_dataclass_equal(
+            result.data,
+            expected_data.data,
+        )
+
+    @pytest.mark.parametrize("sampling,maximum_term", [("64x64", 37), ("128x128", 64)])
     def test_zernike_standard_coefficients_matches_reference_data(
         self, simple_system, sampling, maximum_term, reference_data
     ):
         result = ZernikeStandardCoefficients(sampling=sampling, maximum_term=maximum_term).run(simple_system)
+
+        assert_dataclass_equal(
+            result.data,
+            reference_data.data,
+        )
+
+    @pytest.mark.parametrize("sampling,maximum_term", [("64x64", 37), ("128x128", 64)])
+    def test_zernike_standard_coefficients_decentered_matches_reference_data(
+        self, decentered_system, sampling, maximum_term, reference_data
+    ):
+        result = ZernikeStandardCoefficients(sampling=sampling, maximum_term=maximum_term).run(decentered_system)
 
         assert_dataclass_equal(
             result.data,